% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatipede_example.R
\name{concatipede_example}
\alias{concatipede_example}
\title{Get the path to one of concatipede example files}
\usage{
concatipede_example(example_file = NULL)
}
\arguments{
\item{example_file}{Basename of the target example file. If \code{NULL} (the default), the basenames of the available files are listed.}
}
\value{
The full path to access the example file, or a list of available example files if no \code{example_file} argument was provided.
}
\description{
Several example files are shipped with the concatipede package. This function facilitates the access to those files.
}
\details{
\describe{
\item{COI_Macrobiotidae.fas}{TODO}
\item{ITS2_Macrobiotidae.fas}{TODO}
\item{LSU_Macrobiotidae.fas}{TODO}
\item{SSU_Macrobiotidae.fas}{TODO}
\item{sequences-test-matching.xlsx}{This is an Excel file (extension .xlsx) typically used to test or demonstrate the automatic matching capabilities of the concatipede package. This file represents the Excel template that could be produced by \code{\link{concatipede_prepare}} after detecting the fasta files present in a working directory.}
\item{Macrobiotidae_seqnames.xlsx}{TODO}
}
}
\examples{
concatipede_example()
example <- concatipede_example("sequences-test-matching.xlsx")
if (requireNamespace("readxl")) {
  seqs <- readxl::read_xlsx(example)
  seqs
}

}
