% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata-doc.R
\name{example-files-shipped-with-the-package}
\alias{example-files-shipped-with-the-package}
\title{Example files shipped with the package}
\description{
Several example files are shipped with the concatipede package.
}
\section{Excel file 'sequences-test-matching.xlsx'}{


This is an Excel file (extension .xlsx) typically used to test or
demonstrate the automatic matching capabilities of the concatipede package.

This file represents the Excel template that could be produced by
\code{\link{concatipe_prepare}} after detecting the fasta files present in a
working directory.
}

\examples{
# Excel file 'sequences-test-matching.xlsx'
if (requireNamespace("readxl")) {
  path <- system.file("extdata", "sequences-test-matching.xlsx",
                      package = "concatipede")
  seqs <- read_xlsx(path)
  seqs
}
}
