% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genbank_table.R
\name{get_genbank_table}
\alias{get_genbank_table}
\title{Extract GenBank accession number from correspondence table}
\usage{
get_genbank_table(
  df = NULL,
  filename = NULL,
  writetable = F,
  out = "",
  excel.sheet = 1
)
}
\arguments{
\item{df}{The user-defined correspondence table, as a data frame or equivalent. This is used only if no \code{filename} argument is provided.}

\item{filename}{Filename of input correspondence table. Alternatively, if no filename is provided, the user can provide their own correspondence table as the \code{df} argument.}

\item{writetable}{if TRUE save the Genbank table as excel file in the working directory}

\item{out}{if writetable == T, the name to be attached to the excel filename}

\item{excel.sheet}{specify what sheet from the excel spreadsheet you wanna read. Either a string (the name of a sheet), or an integer (the position of the sheet).}
}
\value{
Table with GenBank accession numbers
}
\description{
Extract GenBank accession number from correspondence table formatted with the same requirements for concatipede()
}
