% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xl.R
\name{read_xl}
\alias{read_xl}
\title{Read an Excel file}
\usage{
read_xl(path, sheet = 1)
}
\arguments{
\item{path}{The path to the Excel file to read.}

\item{sheet}{Optional, the sheet to read (either a string with the name of the sheet or an integer with its position). Default: 1.}
}
\value{
A tibble.
}
\description{
This function loads a table from an Excel file.
}
