% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{reciprocal_matches}
\alias{reciprocal_matches}
\title{Find the best reciprocal matches between x and y}
\usage{
reciprocal_matches(x, y, method = "lv")
}
\arguments{
\item{x, y}{Character vectors. They do not need to have the same length.}

\item{method}{Method for string distance calculation. See
\code{?stringdist::stringdist-metrics}.}
}
\value{
A tibble with columns \code{x} and \code{y} containing the best reciprocal
matches. Strings which are not part of a best reciprocal match are not
returned.
}
\description{
Find the best reciprocal matches between x and y
}
\examples{
x <- c("apple", "banana", "croissant", "dorado", "elephant", "fennec")
y <- c("applet", "ananas", "crescent", "elefant", "fan", "tree", "apply")
reciprocal_matches(x, y)
}
