% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata-doc.R
\name{sequences-test-matching.xlsx}
\alias{sequences-test-matching.xlsx}
\title{Excel file 'sequences-test-matching.xlsx'}
\usage{
system.file("extdata", "sequences-test-matching.xlsx", package = "concatipede")
}
\description{
This is an Excel file (.xlsx) typically used to test or demonstrate the
automatic matching capabilities of the concatipede package.
}
\details{
This file represents the Excel template that could be produced by
\code{\link{concatipe_prepare}} after detecting the fasta files present in a
working directory.
}
\examples{
if (requireNamespace(readxl)) {
  path <- system.file("extdata", "sequences-test-matching.xlsx",
                      package = "concatipede")
  seqs <- read_xlsx(path)
  seqs
}
}
