% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{uuid}
\alias{uuid}
\title{Generate random unique identifiers}
\usage{
uuid(...)
}
\arguments{
\item{...}{All parameters are ignored (but the function does accept them, so
that it can be included in a \code{sapply} call for example).}
}
\value{
A string with 32 characters.
}
\description{
Generate random unique identifiers
}
\examples{
uuid()
sapply(1:10, uuid)
}
