% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aln.R
\name{write_fasta}
\alias{write_fasta}
\title{Save an alignment in fasta format}
\usage{
write_fasta(x, path)
}
\arguments{
\item{x}{Alignment (an object of class \code{DNAbin}).}

\item{path}{Path of the file to be written (the ".fasta" extension is added automatically).}
}
\value{
The input \code{x} (invisibly).
}
\description{
Save an alignment in fasta format
}
