% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aln.R
\name{write_nexus}
\alias{write_nexus}
\title{Save an alignment in nexus format}
\usage{
write_nexus(x, path)
}
\arguments{
\item{x}{Alignment (an object of class \code{DNAbin}).}

\item{path}{Path of the file to be written (the ".nexus" extension is added automatically).}
}
\value{
The input \code{x} (invisibly).
}
\description{
Save an alignment in nexus format
}
