% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aln.R
\name{write_phylip}
\alias{write_phylip}
\title{Save an alignment in phylip format}
\usage{
write_phylip(x, path)
}
\arguments{
\item{x}{Alignment (an object of class \code{DNAbin}).}

\item{path}{Path of the file to be written (the ".phy" extension is added automatically).}
}
\value{
The input \code{x} (invisibly).
}
\description{
Save an alignment in phylip format
}
