% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xl.R
\name{write_xl}
\alias{write_xl}
\title{Write an Excel file}
\usage{
write_xl(x, path)
}
\arguments{
\item{x}{A data frame or tibble to write to an Excel file.}

\item{path}{The path to the Excel file to be written.}
}
\value{
The input table \code{x}, invisibly (so that the function can be part of a pipeline with the pipe operator).
}
\description{
This function writes an input table to an Excel file and returns its input (invisibly).
}
