% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatipede.R
\name{concatipede}
\alias{concatipede}
\title{Concatenate alignments}
\usage{
concatipede(
  filename = "seqnames.txt",
  format = c("fasta", "nexus", "phylip"),
  plotimg = T,
  return.aln = F,
  out = NULL,
  remove.gaps = TRUE,
  write.outputs = TRUE,
  excel.sheet = 1,
  exclude = "concatenated"
)
}
\arguments{
\item{filename}{filename of correspondence table}

\item{format}{a string specifying in what formats you want the alignment}

\item{plotimg}{return a graphical representation of the alignment in pdf format}

\item{return.aln}{return the concatenate alignment inside R workspace}

\item{out}{specify outputs filename}

\item{remove.gaps}{remove gap only columns. Useful if not using all sequences in the alignments}

\item{write.outputs}{save concatenated alignment, partitions position table and graphical representation. If FALSE it overrides plotimg}

\item{excel.sheet}{specify what sheet from the exce spreadsheet you wanna read. Either a string (the name of a sheet), or an integer (the position of the sheet).}

\item{exclude}{fasta files with this text in the working directory will be ingnored by the function}
}
\value{
Can return alignment in the workspace if return.aln is set to TRUE
}
\description{
This function concatenate sequences from alignments present in the working directory based on a correspondence table and saves the output in a new directory
}
