% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatipede.R
\name{concatipede}
\alias{concatipede}
\title{COncatenate alignments}
\usage{
concatipede(
  filename = "seqnames.txt",
  alignments,
  format = c("fasta", "nexus", "phylip"),
  plotimg = T,
  return.aln = F,
  out = NULL
)
}
\arguments{
\item{filename}{filename of correspondence table}

\item{alignments}{output from concatipede_prepare() function}

\item{format}{a string specifying in what formats you want the alignment}

\item{plotimg}{return a graphical representation of the alignment in pdf format}

\item{return.aln}{return the concatenate alignment inside R workspace}

\item{out}{specify outputs filename}
}
\value{
Can return alignment in the workspace if return.aln is set to TRUE
}
\description{
This function concatenate sequences from alignments based on a correspondence table
}
