% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatipede_prepare.R
\name{concatipede_prepare}
\alias{concatipede_prepare}
\title{Load alignments and prepare template correspondence table for concatenate () function}
\usage{
concatipede_prepare(filename = "seqnames", excel = T, exclude = "concatenated")
}
\arguments{
\item{filename}{filename of the saved correspondence table template}

\item{excel}{logic indicating if saving the correspondence table template in the working directory in excel format (otherwise it is saved as text file)}

\item{exclude}{fasta files with this text in the working directory will be ingnored by the function}
}
\value{
List object containing alignments, this object is one of the input for concatipede() function
}
\description{
This function loads all the fasta files present in the working directory,
creates a template correspondence table that is saved in the working directory.
}
