% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genbank_table.R
\name{get_genbank_table}
\alias{get_genbank_table}
\title{Extract GenBank accession number from correspondence table}
\usage{
get_genbank_table(corr_table, writetable = F, out = "")
}
\arguments{
\item{corr_table}{a correlation table. Can be both a dataframe or a character string indicating the file path o the correspondence table in excel format}

\item{writetable}{if TRUE save the Genbank table as excel file in the working directory}

\item{out}{if writetable == T, the name to be attached to the excel filename}
}
\value{
Table with GenBank accession numbers
}
\description{
Extract GenBank accession number from correspondence table formatted with the same requirements for concatipede()
}
